requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.42
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.40.0
numpy==2.2.6
openai==1.79.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
rpds-py==0.25.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
starlette==0.46.2
streamlit==1.45.1
streamlit-cookies-manager==0.2.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5
tqdm==4.67.1
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
